bug(return-type): read should not return null
Normally I would change the return type to add null. However this is a private
function and that option would add 3 more issues as null would be passed to
other php functions that require a resource as input. So I skip the null case
at this function, and maybe another adapter might work better anyhow!